// Front/Frontend/src/app/Models/ReclamationDto.ts
/* tslint:disable */
/* eslint-disable */
export interface ReclamationDto {
    id: number;
    categorie?: string;
    impact?: string;
    lieu?: string;
    urgence?: string;
    employeeMatricule?: string;
    description?: string;
    pieceJointe?: string;
    type?: string;
    status?: string;
    createdAt?: string;
    updatedAt?: string;
    createur?: string;
    // Ajoutez ici d'autres champs spécifiques à la réclamation si nécessaire
  }
  